Refactoring der Struktur des Frontend-Moduls
