+++ frontend/src/main/frontend/src/app/components/activation/activation.component.spec.ts
