+++ frontend/src/main/frontend/src/app/components/alert/alert.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import {AlertService} from "../alert.service";
+import {AlertService} from "../../services/alert.service";
 
 @Component({
   selector: 'app-alert',
