+++ frontend/src/main/frontend/src/app/components/post/post.component.html
