+++ frontend/src/main/frontend/src/app/components/view/view.component.html
