+++ frontend/src/main/frontend/src/app/model/activity-dto.ts
