+++ frontend/src/main/frontend/src/app/model/activity.ts
