+++ frontend/src/main/frontend/src/app/services/activity.service.spec.ts
