+++ frontend/src/main/frontend/src/app/services/activity.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import {HttpClient} from "@angular/common/http";
-import {ActivityDto} from "./activity-dto";
+import {ActivityDto} from "../model/activity-dto";
 import {Observable} from "rxjs/Observable";
-import {Activity} from "./activity";
+import {Activity} from "../model/activity";
 
 @Injectable()
 export class ActivityService {
